+++ rechu/command/__init__.py
@@ -3,6 +3,7 @@ Subcommand collection package.
 """
 
 from .base import Base
+from .alembic import Alembic
 from .create import Create
 from .new import New
 from .read import Read