+++ CHANGELOG.md
@@ -9,5 +9,6 @@ and we adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Database migration support using Alembic through a subcommand added.
 - Initial version with database schema creation and YAML file reading/writing 
   for database import.